Create reusable CardTooltip atom
