+++ frontend/src/components/ReduceGraphSwitch.tsx
@@ -18,13 +18,9 @@ import * as React from 'react';
 import Switch, { SwitchProps } from '@material-ui/core/Switch';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Tooltip from '@material-ui/core/Tooltip';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import { withStyles } from '@material-ui/core/styles';
 import { stylesheet } from 'typestyle';
-import { color, fontsize } from '../Css';
 import { ExternalLink } from '../atoms/ExternalLink';
+import { CardTooltip } from '../atoms/CardTooltip';
 
 const css = stylesheet({
   reduceGraphPosition: {
@@ -34,47 +30,29 @@ const css = stylesheet({
   },
 });
 
-const NostyleTooltip = withStyles({
-  tooltip: {
-    backgroundColor: 'transparent',
-    border: '0 none',
-    color: color.secondaryText,
-    fontSize: fontsize.base,
-    maxWidth: 220,
-  },
-})(Tooltip);
-
 interface ReduceGraphSwitchProps extends SwitchProps {}
 
 const ReduceGraphSwitch = (props: ReduceGraphSwitchProps) => {
   return (
     <div className={css.reduceGraphPosition}>
-      <NostyleTooltip
-        title={
-          <Card>
-            <CardContent>
-              <div>
-                <p>
-                  Enables a{' '}
-                  <ExternalLink href='https://en.wikipedia.org/wiki/Transitive_reduction'>
-                    transitive reduction
-                  </ExternalLink>{' '}
-                  of the pipeline graph, hiding all the redundant edges. This option is just a
-                  visualization helper and does not have any permanent effect on the pipeline
-                  itself.
-                </p>
-                <p>
-                  Note that edges are removed regardless of their function in the pipeline, so the
-                  reduced form might not provide enough details to understand how the pipeline
-                  really behaves.
-                </p>
-              </div>
-            </CardContent>
-          </Card>
+      <CardTooltip
+        helpText={
+          <div>
+            <p>
+              Enables a{' '}
+              <ExternalLink href='https://en.wikipedia.org/wiki/Transitive_reduction'>
+                transitive reduction
+              </ExternalLink>{' '}
+              of the pipeline graph, hiding all the redundant edges. This option is just a
+              visualization helper and does not have any permanent effect on the pipeline itself.
+            </p>
+            <p>
+              Note that edges are removed regardless of their function in the pipeline, so the
+              reduced form might not provide enough details to understand how the pipeline really
+              behaves.
+            </p>
+          </div>
         }
-        interactive={true}
-        leaveDelay={200}
-        placement='top'
       >
         <FormGroup>
           <FormControlLabel
@@ -83,7 +61,7 @@ const ReduceGraphSwitch = (props: ReduceGraphSwitchProps) => {
             labelPlacement='end'
           />
         </FormGroup>
-      </NostyleTooltip>
+      </CardTooltip>
     </div>
   );
 };
